fea:create auth context.
